+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 	## __WORK IN PROGRESS__
 -->
 
+## __WORK IN PROGRESS__
+* Unit tests for adapter startup were removed and only log a warning that you can remove them
+
 ## 2.3.0 (2020-08-20)
 * Added missing async functions to adapter mock
 * Fixed: `TypeError "Cannot redefine property readyHandler"` when using `createMocks` more than once